Support searching the English Wikipedia before fetching descriptions
Add a new popups gateway that wraps around the existing rest gateway, but first
makes an API call to the English Wikipedia's `opensearch` endpoint for the
requested title. This interaction is handled via the `mw.ForeignApi` class
provided by MediaWiki core, necessitating the addition of the
`mediawiki.ForeignApi` ResourceLoader module as a dependency of the popups
module. If there are any results from that API call, the popup is rendered for
the first result. If there are no results, no popup is render, but no error is
shown about the failure to fetch a popup.

The `opensearch` API call is only done once per link, with the results stored
in the `opensearch-results` attribute of the link, where an empty string
indicates no results and any other string indicates the name of the page that
should be fetched.

The new popup gateway is enabled via `$wgPopupsGateway = 'restbaseWithSearch';`

By default, when dwelling on a link that the popups extension is managing, the
default `title` attribute is removed until the user abandons the link, so that
it does not conflict with the actual popup. Once it has been established that a
link should not have a popup, however, this behavior is skipped, and the
original title text is shown on future occasions when the user dwells on the
link. The original `title` attribute is also added back immediately for the
current dwelling, however, at least on Chrome the restored `title` is not
shown, likely because it needs to be present from the start to be shown. The
behavior can be confirmed via inspecting the DOM.

Additionally, support having popups for pages that are "redlinked", i.e. that
do not exist locally. This support is unconditional, as our fork of the
extension is expected to only be used with the new gateway introduced. To add
this support

* Stop excluding links with the `new` class from the set of links that have the
popups event handlers attached.

* When analyzing the query parameters of a link, if the parameters include
`action=edit` and `redlink=1`, delete both of those, since otherwise the
presence of query parameters were prevent the popups extension from handling
the links.

Also modify `.gitattributes` to indicate that all text should end with LF, to
suppress some annoying warnings.

SEL-1387