restbaseWithSearch: link to the English Wikipedia
When using the custom `restbaseWithSearch` gateway that will always show popups
from the English Wikipedia, modify the returned page models to link to the
pages on the English Wikipedia rather than pages on the local wiki.

Additionally, update the popups so that the links open in a new tab.

SEL-1348